# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Book Lending API Backend
  description: An API backend for a book lending service
  contact:
    name: Joe Standring
    email: git@joestandring.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
servers:
  - url: 'https://mixer-fiona-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /requests:
    summary:
      Represents an individual request for a book.
    description: >
      This resource represents a request for a book in the system.
      Each request is represented by a numeric ID.
    get:
      tags:
        - Requests
      description:
        Get all user requests for books.
      responses:
        '200':
          description:
            View of all book requests.
        '401':
          description:
            Non-admin users cannot see all book requests.
    post:
      tags:
        - Requests
      description:
        Create a request for a book.
      responses:
        '201':
          description:
            Request was successfully made for book.
        '401':
          description:
            You must be signed in to make book requests.
  /requests/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description:
          Numeric ID of a book request.
    get:
      tags:
        - Requests
      description:
        Get a book request by ID.
      responses:
        '200':
          description:
            View of a single book request.
        '401':
          description:
            Non-admin users can only view their own book requests.
        '404':
          description:
            Request for book was not found.
    put:
      tags:
        - Requests
      description:
        Edit a single book request by ID.
      responses:
        '201':
          description:
            Book request was successfully updated.
        '400':
          description:
            Improperly formatted request.
        '401':
          description:
            Non-admin users can only update their own book requests.
        '404':
          description:
            Book request was not found.
    delete:
      tags:
        - Requests
      description:
        Delete a single book by ID.
      responses:
        '200':
          description:
            Book request was successfully deleted.
        '401':
          description:
            Non-admin users can only delete their own book requests.
        '404':
          description:
            Book request was not found.
...